+++ Runtime/SdkVersion.cs
@@ -10,6 +10,6 @@ namespace Unity.Services.Analytics
         /// This value should always match what is in the package.json file for this package,
         /// to provide accurate SDK versions to the analytics data classes.
         /// </summary>
-        public static readonly string SDK_VERSION = "6.0.3";
+        public static readonly string SDK_VERSION = "6.1.0";
     }
 }